* Most changes are to fix warnings issued when compiling win32 * removed a few redundant defines * get_user_name safe under win32 * rationalized pipe read EOF for win32 (UPDATED PATCH USED) * changed all backend instances of sleep() to pg_usleep
    - except for the SLEEP_ON_ASSERT in assert.c, as it would exceed a
32-bit long [Note to patcher: If a SLEEP_ON_ASSERT of 2000 seconds is
acceptable, please replace with pg_usleep(2000000000L)]

I added a comment to that part of the code:

    /*
     *  It would be nice to use pg_usleep() here, but only does 2000 sec
     *  or 33 minutes, which seems too short.
     */
    sleep(1000000);

Claudio Natoli

+++ src/backend/port/win32/timer.c
@@ -6,7 +6,7 @@
  * Portions Copyright (c) 1996-2003, PostgreSQL Global Development Group
  *
  * IDENTIFICATION
- *	  $PostgreSQL: pgsql/src/backend/port/win32/timer.c,v 1.1 2004/02/18 16:25:12 momjian Exp $
+ *	  $PostgreSQL: pgsql/src/backend/port/win32/timer.c,v 1.2 2004/04/19 17:42:58 momjian Exp $
  *
  *-------------------------------------------------------------------------
  */
@@ -43,9 +43,9 @@ int setitimer(int which, const struct itimerval *value, struct itimerval *ovalue
 		timerHandle = CreateWaitableTimer(NULL, TRUE, NULL);
 		if (timerHandle == NULL)
 			ereport(FATAL,
-					(errmsg_internal("failed to create waitable timer: %i",GetLastError())));
+					(errmsg_internal("failed to create waitable timer: %i",(int)GetLastError())));
 	}
-	
+
 	if (value->it_value.tv_sec == 0 &&
 		value->it_value.tv_usec == 0) {
 		/* Turn timer off */
@@ -55,11 +55,11 @@ int setitimer(int which, const struct itimerval *value, struct itimerval *ovalue
 
 	/* Negative time to SetWaitableTimer means relative time */
 	dueTime.QuadPart = -(value->it_value.tv_usec*10 + value->it_value.tv_sec*10000000L);
-	
+
 	/* Turn timer on, or change timer */
-	if (!SetWaitableTimer(timerHandle, &dueTime, 0, timer_completion, NULL, FALSE)) 
+	if (!SetWaitableTimer(timerHandle, &dueTime, 0, timer_completion, NULL, FALSE))
 		ereport(FATAL,
-				(errmsg_internal("failed to set waitable timer: %i",GetLastError())));
+				(errmsg_internal("failed to set waitable timer: %i",(int)GetLastError())));
 
 	return 0;
 }
